split DocumentEditor/heading into shared/ui
