+++ packages/fields-document/src/DocumentEditor/index.tsx
@@ -29,7 +29,7 @@ import { withLayouts } from './layouts'
 import { clearFormatting, type Mark } from './utils'
 import { Toolbar } from './Toolbar'
 import { renderElement } from './render-element'
-import { withHeading } from './heading'
+import { withHeading } from './heading-shared'
 import { nestList, unnestList, withList } from './lists-shared'
 import { ComponentBlockContext, withComponentBlocks } from './component-blocks'
 import { getPlaceholderTextForPropPath } from './component-blocks/utils'
